Heavy revision of gcomp_boot() and add_gcomputation() functions. They now allow both continuous and dichotomous outcome variables called in the gtsummary tables, while also handling different variables in the same output.
Now possible to calculate both average and individual treatment estimates through G-computation, with both marginal risk differences, risk ratios, mean differences and ratio of means for both adjusted and unadjusted analysis setup, including handling of footnotes for both combined and separate footnotes in the final output.

Few tests updated, as well as namespace, description and the readme for the package.